+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Fixed messages being sent to the wrong contact ([#615])
+- Fixed incomplete message exports ([#713])
+- Fixed crash when viewing older messages
 
 ## [1.8.0] - 2026-01-30
 ### Added
@@ -235,9 +239,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#574]: https://github.com/FossifyOrg/Messages/issues/574
 [#600]: https://github.com/FossifyOrg/Messages/issues/600
 [#610]: https://github.com/FossifyOrg/Messages/issues/610
+[#615]: https://github.com/FossifyOrg/Messages/issues/615
 [#641]: https://github.com/FossifyOrg/Messages/issues/641
 [#644]: https://github.com/FossifyOrg/Messages/issues/644
 [#651]: https://github.com/FossifyOrg/Messages/issues/651
+[#713]: https://github.com/FossifyOrg/Messages/issues/713
 
 [Unreleased]: https://github.com/FossifyOrg/Messages/compare/1.8.0...HEAD
 [1.8.0]: https://github.com/FossifyOrg/Messages/compare/1.7.0...1.8.0